+++ client/src/extension.ts
@@ -44,6 +44,7 @@ import getport from "get-port";
 import execa from "execa";
 import { init, localize } from "vscode-nls-i18n";
 import { MockDebugSession } from "./mockDebug";
+import * as semver from "semver";
 
 import { TreeViewProvider } from "./tree_view_provider";
 
@@ -690,14 +691,14 @@ Executable ${this.denoInfo.executablePath}`;
             cancellable: true,
           },
           (process, cancelToken) => {
-            const ps = execa(
-              this.denoInfo.executablePath,
-              ["fetch", moduleName],
-              {
-                // timeout of 2 minute
-                timeout: 1000 * 60 * 2,
-              }
-            );
+            // `deno fetch xxx` has been renamed to `deno cache xxx` since Deno v0.40.0
+            const cmd = semver.gte(this.denoInfo.version.deno, "0.40.0")
+              ? "cache"
+              : "fetch";
+            const ps = execa(this.denoInfo.executablePath, [cmd, moduleName], {
+              // timeout of 2 minute
+              timeout: 1000 * 60 * 2,
+            });
 
             const updateProgress = (buf: Buffer) => {
               const raw = buf.toString();