+++ core/import_map.ts
@@ -18,7 +18,10 @@ type ImportFileMapContent = {
 type ImportContent = { [prefix: string]: string };
 
 export class ImportMap implements ImportMapInterface {
-  constructor(public map: ImportFileMapContent, public filepath?: string) {}
+  private constructor(
+    public map: ImportFileMapContent,
+    public filepath?: string
+  ) {}
   static create(importMapFilepath?: string): ImportMapInterface {
     let importMap: ImportFileMapContent = {
       imports: {},
@@ -39,11 +42,31 @@ export class ImportMap implements ImportMapInterface {
 
         try {
           importMap = JSON.parse(importMapContent || "");
+
+          // Make sure `importMap.imports` is a key-value object
+          // Otherwise, an exception may be thrown
+          if (
+            Object.prototype.toString.call(importMap.imports) !==
+            "[object Object]"
+          ) {
+            importMap.imports = {};
+          }
         } catch {
           importMap.imports = {};
         }
       }
     }
+
+    const imports = Object.entries(importMap.imports)
+      .filter(
+        ([key, val]) =>
+          (key.endsWith("/") && val.endsWith("/")) ||
+          (!key.endsWith("/") && !val.endsWith("/"))
+      )
+      .sort(([key1], [key2]) => key2.lastIndexOf("/") - key1.lastIndexOf("/"))
+      .reduce((imports, [key, value]) => ({ ...imports, [key]: value }), {});
+
+    importMap.imports = imports;
 
     return new ImportMap(importMap, importMapFilepath);
   }